Rename "identifer" app
