+++ gear/identifier_lookup/src/docker/BUILD
@@ -1,9 +1,9 @@
 file(name="manifest", source="manifest.json")
 
-docker_image(name="identifer-lookup",
+docker_image(name="identifier-lookup",
              source="Dockerfile",
              dependencies=[
                  ":manifest",
-                 "gear/identifer_lookup/src/python/identifer_app:bin"
+                 "gear/identifier_lookup/src/python/identifier_app:bin"
              ],
              image_tags=["0.0.1", "latest"])
