+++ gear/identifier_lookup/src/docker/Dockerfile
@@ -1,0 +1,11 @@
+FROM python:3.10
+
+ENV BASE_DIR=/flywheel/v0
+RUN mkdir -p ${BASE_DIR}/input
+
+WORKDIR ${BASE_DIR}
+
+COPY gear/identifier_lookup/src/docker/manifest.json ${BASE_DIR}
+COPY gear.identifier_lookup.src.python.${APPNAME}/bin.pex /bin/run
+
+ENTRYPOINT [ "/bin/run" ]
