+++ gear/identifier_lookup/src/python/identifier_app/BUILD
@@ -1,3 +1,3 @@
-python_sources(name="identifer_app", )
+python_sources(name="identifier_app", )
 
 pex_binary(name="bin", entry_point="run.py")
